requirements.txt
# ML Libraries
torch
datasets
transformers
tokenizers
sentence-transformers

# Training and Evaulation
scikit-learn
accelerate
evaluate
umap-learn

# Data Processing
pandas
openpyxl
numpy

# Visualization
matplotlib
plotly
seaborn


# API Development
fastapi
uvicorn
pydantic
streamlit
plotly

#Utilties
tqdm
requests
overrides
python-dotenv

numba==0.59.1
llvmlite==0.42.0
umap-learn